+++ rollup.config.mjs
@@ -4,7 +4,9 @@ import resolve from '@rollup/plugin-node-resolve';
 import { terser } from 'rollup-plugin-terser';
 import { sveltePreprocess } from 'svelte-preprocess';
 import typescript from '@rollup/plugin-typescript';
-import css from 'rollup-plugin-css-only';
+import postcss from 'rollup-plugin-postcss';
+import tailwindcss from 'tailwindcss';
+import tailwindConfig from './tailwind.config.js';
 
 const production = !process.env.ROLLUP_WATCH;
 
@@ -26,7 +28,16 @@ export default {
 		}),
 		// we'll extract any component CSS out into
 		// a separate file - better for performance
-		css({ output: 'bundle.css' }),
+		postcss({
+			config: {
+			  path: './postcss.config.js',
+			},
+			extensions: [".css"],
+			minimize: true,
+			extract: true,
+			plugins: [tailwindcss(tailwindConfig)],
+			output: "bundle.css"
+		}),
 
 		// If you have external dependencies installed from
 		// npm, you'll most likely need these plugins. In